Rewrite RadioGroupField
